Introduce "one click edition" feature.
Now that the documentation has been moved to github
we can add a link above each documentation page that
allows people to directly edit the page on github. When
the page is saved, github automatically creates a new
pull request with the changes.

We should wait that googlebot is configured for our
documentation repository before to merge this change.

Change-Id: I4d20932d4ec05e86b65d25a71c0c51ac03a39e76